feat: Component cutout of appbar
