+++ app/src/components/modules/Dashboard.tsx
@@ -1,28 +1,11 @@
-import React from 'react'
-import clsx from 'clsx'
 import { makeStyles } from '@material-ui/core/styles'
-import { useNavigate } from 'react-router-dom'
-import AppBar from '@material-ui/core/AppBar'
 import Box from '@material-ui/core/Box'
 import CssBaseline from '@material-ui/core/CssBaseline'
-import Drawer from '@material-ui/core/Drawer'
-import Toolbar from '@material-ui/core/Toolbar'
-import List from '@material-ui/core/List'
 import Typography from '@material-ui/core/Typography'
-import Divider from '@material-ui/core/Divider'
-import IconButton from '@material-ui/core/IconButton'
-import Badge from '@material-ui/core/Badge'
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft'
 import Container from '@material-ui/core/Container'
 import Link from '@material-ui/core/Link'
-import LogoutIcon from '@mui/icons-material/Logout'
-import Menu from '@mui/material/Menu'
-import MenuIcon from '@material-ui/icons/Menu'
-import MenuItem from '@mui/material/MenuItem'
-import NotificationsIcon from '@material-ui/icons/Notifications'
-import SettingsIcon from '@mui/icons-material/Settings'
 // My modules
-import { mainListItems, secondaryListItems } from './listItems'
+import DashboardBar from './DashboardBar'
 
 function MainView(props: any) {
   return (
@@ -128,113 +111,12 @@ const useStyles = makeStyles((theme) => ({
 
 export default function Dashboard(props: any) {
   const classes = useStyles()
-  const navigate = useNavigate()
   const element = <MainView name={props.name} />
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
-
-  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget)
-  }
-
-  const handleClose = () => {
-    setAnchorEl(null)
-  }
-
-  const logOut = () => {
-    /**
-     * TODO: After log out,  /user-regist/.
-     *  Specify the root path
-     */
-    document.cookie = 'auth_token=; path=/'
-    navigate('/')
-  }
-
-  const [open, setOpen] = React.useState(true)
-  const handleDrawerOpen = () => {
-    setOpen(true)
-  }
-  const handleDrawerClose = () => {
-    setOpen(false)
-  }
-  const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight)
 
   return (
     <div className={classes.root}>
       <CssBaseline />
-      <AppBar
-        position="absolute"
-        className={clsx(classes.appBar, open && classes.appBarShift)}
-      >
-        <Toolbar className={classes.toolbar}>
-          <IconButton
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            className={clsx(
-              classes.menuButton,
-              open && classes.menuButtonHidden
-            )}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography
-            component="h1"
-            variant="h6"
-            color="inherit"
-            noWrap
-            className={classes.title}
-          >
-            Dashboard
-          </Typography>
-          <IconButton color="inherit">
-            <Badge badgeContent={5} color="secondary">
-              <NotificationsIcon />
-            </Badge>
-          </IconButton>
-          <IconButton color="inherit" onClick={handleMenu}>
-            <SettingsIcon />
-          </IconButton>
-          <Menu
-            id="menu-appbar"
-            anchorEl={anchorEl}
-            anchorOrigin={{
-              vertical: 'top',
-              horizontal: 'right',
-            }}
-            keepMounted
-            transformOrigin={{
-              vertical: 'top',
-              horizontal: 'right',
-            }}
-            open={Boolean(anchorEl)}
-            onClose={handleClose}
-          >
-            {/* TODO: Log Out のみ実装した */}
-            <MenuItem onClick={handleClose}>todo: Profile</MenuItem>
-            <MenuItem onClick={handleClose}>todo: My account</MenuItem>
-            <MenuItem onClick={logOut}>
-              <LogoutIcon /> Logout
-            </MenuItem>
-          </Menu>
-        </Toolbar>
-      </AppBar>
-      <Drawer
-        variant="permanent"
-        classes={{
-          paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
-        }}
-        open={open}
-      >
-        <div className={classes.toolbarIcon}>
-          <IconButton onClick={handleDrawerClose}>
-            <ChevronLeftIcon />
-          </IconButton>
-        </div>
-        <Divider />
-        <List>{mainListItems}</List>
-        <Divider />
-      </Drawer>
+      <DashboardBar />
       <main className={classes.content}>
         {/* ここに各種ページ内容を埋め込む */}
         <Container maxWidth="lg" className={classes.container}>
